Added controller models and VR-compatible stats
Pulled in some temporary controller models while I wait for the WebXR
input profiles repo to be populated.

Also ported my WebGL-based FPS counter to Three.js for use here. This
both allows the FPS to be displayed in VR (attached to the back of the
right hand controller) and avoids mixing a 2D canvas with WebGL as was
the case for the standard stats lib. Turns out this was causing a
performance issue in Chrome 79 (which is being investagated.)