Make use of THREE.Curve to generate better guideline geometry
Not using TubeBufferGeometry here because the cross shape we have now works better for the effect that I want and it's not clear how you'd update the TubeBufferGeometry in place. Seems like you'd have to allocate a whole new one every frame. Fortunately the THREE.Curve class gives us the primary thing we need, which is the FrenetFrames.